+++ models/inventory.go
@@ -5,7 +5,7 @@ import "time"
 type InventoryItem struct {
     IngredientID int64  
     Name         string  
-    Quantity     float64
+    Quantity     int
     Unit         string  
 	CreatedAt    time.Time
 	UpdatedAt    time.Time
@@ -18,7 +18,7 @@ type InventoryItem struct {
 // }
 
 // NewInventoryItem создаёт новый InventoryItem с уникальным int64 ID
-func NewInventoryItem(name string, quantity float64, unit string) InventoryItem {
+func NewInventoryItem(name string, quantity int, unit string) InventoryItem {
     return InventoryItem{
         IngredientID: 0, // заполняется после вставки в БД
         Name:         name,
